requirements.txt
sounddevice
paho-mqtt
asyncio-paho
torch
transformers
deep-translator
pygame
gTTS
pyttsx3

